projects: max14906 : Change interrupt support for MAXIM platform
Move specific interrupt support related function calls in
maxim main source file.
Add error labels in maxim main source file to handle the errors
that may appear from the new function calls.
